moved CategoryDropdown into a separate component
